created authentication using jwt's and alternating login/logout buttons based on token existing
