+++ client/src/pages/Home.js
@@ -6,7 +6,7 @@ import TextField from '@mui/material/TextField';
 
 // API
 import { RepList } from "../components/RepList";
-import { VoterInfo } from "../components/VoterInfo";
+import { VoterInfo } from "../components/voterInfo";
 
 const Home = () => {
     return (
