+++ server/schema/resolvers.js
@@ -23,11 +23,10 @@ const resolvers = {
 
     Mutation: {
         addUser: async(parent, args) => {
-
-            console.log(args)
             
             const user = await User.create(args);
             const token = signToken(user);
+            console.log(token);
 
             return {token, user};
         },
